Add linter and start static page functionality
