+++ dpi_calculator_page/style.css
@@ -1,0 +1,70 @@
+body {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  font-family: 'Times New Roman', Times, serif;
+  background-color: darkslategrey;
+  color: cornsilk;
+}
+
+h1 {
+  font-size: 2.5em;
+  margin-bottom: 0.5em;
+  position: relative;
+  text-align: center;
+}
+
+.btn {
+  background-color: darkslategray;
+  color: cornsilk;
+  padding: 0.5em 1em;
+  border: 2px solid cornsilk;
+  border-radius: 5px;
+  cursor: pointer;
+  margin: 0.5em;
+}
+
+#content-wrapper {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  margin: auto;
+  flex: '1 1 100%';
+  overflow-y: auto;
+}
+
+#content {
+  text-align: center;
+  margin: auto;
+}
+
+#boundary_box {
+  width: 900px;
+  margin: 0 auto;
+  align-items: center; 
+  justify-content: center;
+  display: flex;
+  flex-direction: column;
+}
+
+#resize_element {
+  width: 100%;
+  height: 100%;
+  border: 2px solid black;
+  border-radius: 10px;
+  margin: 5px auto;
+  background-color: aquamarine;
+  position: relative;
+}
+
+#drag_element {
+  position: absolute;
+  cursor: nwse-resize;
+  bottom: 0;
+  right: 0;
+  background-color: black;
+  border-radius: 5px;
+}
