+++ os_resizing_plugin/os_resizing.ts
@@ -1,30 +1,3 @@
-/**
- * The `Resizer` class provides functionality to create a resizable element on a canvas.
- * It initializes with default dimensions for a card and a resize element, and maintains
- * the aspect ratio during resizing.
- *
- * @remarks
- * The class creates a content div element, a resize element, and a button to trigger the resize.
- * It also handles mouse events to allow dragging and resizing of the element.
- *
- * @example
- * ```typescript
- * const resizer = new Resizer();
- * ```
- *
- * @property {number} init_height - The initial height of the card in millimeters.
- * @property {number} init_width - The initial width of the card in millimeters.
- * @property {number} init_resize_element - The initial size of the resize element in pixels.
- * @property {number} aspect_ratio - The aspect ratio of the card (width/height).
- * @property {number} px2mm - The conversion factor from pixels to millimeters.
- *
- * @method test_div - Creates and styles a content div element, and appends it to the canvas parent.
- * @method force_canvas_size - Forces the content div element to match the canvas size.
- * @method create_resize_element - Creates a resizable element and appends it to the content div.
- * @method create_drag_element - Creates a draggable element for resizing and appends it to the resize element.
- * @method create_btn - Creates a button to trigger the resize and appends it to the content div.
- * @method resize_object - Handles the resizing logic and updates the canvas size based on the resize element.
- */
 class Resizer {
     private init_height: number = 53.98;
     private init_width: number = 85.6;
@@ -51,147 +24,144 @@ class Resizer {
     scale_factor: number;
     use_perceived_distance: boolean;
 
+    constructor(osweb: boolean, runner: any, use_perceived_distance: boolean, development_dpi: number, development_distance: number) {
+        if (!osweb) {
+            this.static_page_main();
+            return;
+        }
 
-    constructor(runner: any, use_perceived_distance: boolean, development_dpi: number, development_distance: number) { // TODO: add DPI and bool for perceived distance
         this.development_dpi = development_dpi;
         this.development_distance = development_distance;
         this.use_perceived_distance = use_perceived_distance;
         this.runner = runner;
-        this.main();
+        this.osweb_main();
     }
 
-    private main() {
-        var content_wrapper = this.create_content_wrapper();
+    private static_page_main() {
+
+    }
+
+    private osweb_main() {
+        const content_wrapper = this.create_content_wrapper();
         this.content_div(content_wrapper);
         this.resize_object();
         this.get_keyboard_response = this.get_keyboard_response.bind(this);
     }
 
-    /**
-     * Creates and styles a content wrapper div element, and appends it to the body.
-     * @returns {HTMLElement} The content wrapper div element.
-     */
     create_content_wrapper() {
         const content_wrapper = document.createElement('div');
         content_wrapper.id = 'content-wrapper';
-        content_wrapper.style.display = 'flex';
-        content_wrapper.style.flexDirection = 'column';
-        content_wrapper.style.margin = 'auto';
-        content_wrapper.style.width = '100%';
-        content_wrapper.style.flex = '1 1 100%';
-        content_wrapper.style.overflowY = 'auto';
+        add_style();
+
         document.body.appendChild(content_wrapper);
         return content_wrapper;
+
+        function add_style() {
+            content_wrapper.style.display = 'flex';
+            content_wrapper.style.flexDirection = 'column';
+            content_wrapper.style.justifyContent = 'center';
+            content_wrapper.style.alignItems = 'center';
+            content_wrapper.style.width = '100%';
+            content_wrapper.style.margin = 'auto';
+            content_wrapper.style.flex = '1 1 100%';
+            content_wrapper.style.overflowY = 'auto';
+        }
     }
 
-    /**
-     * Creates and styles a content div element, and appends it to the canvas parent.
-     * @returns {void}
-     */
     private content_div(content_wrapper: HTMLElement): void {
-        var content = document.createElement('div');
+        document.body.getElementsByTagName('main')[0].style.display = 'none';
+
+        let content = document.createElement('div');
         content.id = 'content';
         content.style.textAlign = 'center';
         content.style.margin = 'auto';
 
-        document.body.getElementsByTagName('main')[0].style.display = 'none';
+        const boundary_box = document.createElement('div');
+        boundary_box.id = 'boundary_box';
+        boundary_box.style.width = '900px';
+        boundary_box.style.margin = '0 auto';
 
-        let insert_name = document.createElement('div');
-        insert_name.id = 'insert_name';
-        insert_name.style.width = '900px';
-        insert_name.style.margin = '0 auto';
-
-        this.create_resize_element(insert_name);
-        this.create_btn(insert_name);
-        content.appendChild(insert_name);
+        this.create_resize_element(boundary_box);
+        this.create_btn(boundary_box);
+        content.appendChild(boundary_box);
         content_wrapper.appendChild(content);
     }
 
-    /**
-     * Creates the resizing element and appends it to the content div.
-     * @param content content div element that will contain the resize element
-     */
-    private create_resize_element(insert_name: HTMLElement): void {
-        let page_size = document.createElement('div');
-        page_size.id = 'page_size';
-        insert_name.appendChild(page_size);
+    private create_resize_element(boundary_box: HTMLElement): void {
         this.aspect_ratio = this.init_width / this.init_height;
-        let resize_element = document.createElement('div');
+        const resize_element = document.createElement('div');
         resize_element.id = 'resize_element';
-        let start_div_height = 
-            this.aspect_ratio < 1 ? this.init_resize_element : Math.round(this.init_resize_element / this.aspect_ratio); // aspect ratio < 1 means width < height
-        let start_div_width = 
-            this.aspect_ratio < 1 ? Math.round(this.init_resize_element / this.aspect_ratio) : this.init_resize_element;
-        var adjust_size = Math.round(start_div_width * 0.1);
+        const start_div_height = 
+        this.aspect_ratio < 1 ? this.init_resize_element : Math.round(this.init_resize_element / this.aspect_ratio); // aspect ratio < 1 means width < height
+        const start_div_width = 
+        this.aspect_ratio < 1 ? Math.round(this.init_resize_element / this.aspect_ratio) : this.init_resize_element;
+        let adjust_size = Math.round(start_div_width * 0.1);
         
-        resize_element.style.border = 'none';
-        resize_element.style.height = start_div_height + 'px';
-        resize_element.style.width = start_div_width + 'px';
-        resize_element.style.margin = '5px auto';
-        resize_element.style.background = '#006600';
-        resize_element.style.position = 'relative';
-
+        add_style();
+        
         this.create_drag_element(resize_element, adjust_size);
-        page_size.appendChild(resize_element);
+        boundary_box.appendChild(resize_element);
+
+        function add_style() {
+            resize_element.style.border = '2px solid black';
+            resize_element.style.height = start_div_height + 'px';
+            resize_element.style.width = start_div_width + 'px';
+            resize_element.style.margin = '5px auto';
+            resize_element.style.background = 'aquamarine';
+            resize_element.style.position = 'relative';
+        }
     }
 
-    /**
-     * Creates a draggable element for resizing and appends it to the resize element.
-     * @param resize_element Resize element that will contain the drag element
-     * @param adjust_size The size of the drag element in pixels
-     */
     private create_drag_element(resize_element: HTMLElement, adjust_size: number) {
-        let drag_element = document.createElement('div');
+        const drag_element = document.createElement('div');
         drag_element.id = 'drag_element';
-        drag_element.style.position = 'absolute';
-        drag_element.style.cursor = 'nwse-resize';
-        drag_element.style.bottom = '0';
-        drag_element.style.right = '0';
-        drag_element.style.width = adjust_size + 'px';
-        drag_element.style.height = adjust_size + 'px';
-        drag_element.style.border = '1px solid red';
-        drag_element.style.backgroundColor = 'none';
-        drag_element.style.borderLeft = '0';
-        drag_element.style.borderTop = '0';
-        drag_element.style.cursor = 'move';
+        add_style();
         resize_element.appendChild(drag_element);
+
+        function add_style() {
+            drag_element.style.position = 'absolute';
+            drag_element.style.cursor = 'nwse-resize';
+            drag_element.style.bottom = '0';
+            drag_element.style.right = '0';
+            drag_element.style.width = adjust_size + 'px';
+            drag_element.style.height = adjust_size + 'px';
+            drag_element.style.border = '1px solid red';
+            drag_element.style.backgroundColor = 'none';
+            drag_element.style.borderLeft = '0';
+            drag_element.style.borderTop = '0';
+            drag_element.style.cursor = 'move';
+        }
     }
 
-    /**
-     * Creates a button to trigger the resize and appends it to the content div.
-     * @param content content div element that contains the resize element
-     */
-    private create_btn(insert_name: HTMLElement) {
-        let btn = document.createElement('button');
+    private create_btn(boundary_box: HTMLElement) {
+        const btn = document.createElement('button');
         btn.id = 'resize_btn';
-        btn.textContent = 'Resize';
+        add_style();
 
-        btn.style.display = 'inline-block';
-        btn.style.margin = '0.75em';
-        btn.style.textAlign = 'center';
-        btn.style.verticalAlign = 'middle';
-        btn.style.position = 'relative';
-        btn.style.bottom = '0';
+        boundary_box.appendChild(btn);
 
-        insert_name.appendChild(btn);
+        function add_style() {
+            btn.textContent = 'Resize';
+            btn.style.display = 'inline-block';
+            btn.style.margin = '0.75em';
+            btn.style.textAlign = 'center';
+            btn.style.verticalAlign = 'middle';
+            btn.style.position = 'relative';
+            btn.style.bottom = '0';
+        }
     }
 
-    /**
-     * Handles the resizing logic and updates the canvas size based on the resize element.
-     * @returns {void}
-     * @throws {Error} Resize element not found
-     */
-    resize_object() {
+    resize_object() { // TODO: add variable to make static page possible without blindspot task
         this._complete_function_cache = this.runner._events._currentItem._complete; // cache the complete function
         this.runner._events._currentItem._complete = () => {}; // override the complete function
         let dragging = false;
-        let resize_element = document.querySelector<HTMLElement>('#resize_element');
+        const resize_element = document.querySelector<HTMLElement>('#resize_element');
         if (!resize_element) {
             throw new Error('Resize element not found');
         };
 
-        let original_height = parseInt(resize_element.style.height);
-        let original_width = parseInt(resize_element.style.width);
+        const original_height = parseInt(resize_element.style.height);
+        const original_width = parseInt(resize_element.style.width);
         let origin_x, origin_y;
 
         document.addEventListener('mouseup', () => {
@@ -230,11 +200,8 @@ class Resizer {
         });
     }
 
-    /**
-     * Starts the blindspot task.
-     */
     start_blindspot_task() {
-        let div = document.querySelector<HTMLElement>('#insert_name');
+        let div = document.querySelector<HTMLElement>('#boundary_box');
         if (!div) {
             throw new Error('Test div not found');
         }
@@ -282,14 +249,6 @@ class Resizer {
         this.reset_ball_wait_for_start();
     }
 
-    /**
-     * Get a keyboard response.
-     * @param callback_function function to be triggered on response 
-     * @param valid_responses array of valid responses
-     * @param persist whether to keep the listener after a valid response
-     * @param allow_held_keys allow held keys
-     * @param minimum_rt set a minimum response time
-     */
     get_keyboard_response(
         callback_function: (response: { key: string, rt: number }) => void,
         valid_responses: string[],
@@ -316,9 +275,6 @@ class Resizer {
         this.listeners.push(listener);
     }
 
-    /**
-     * Start the ball animation.
-     */
     start_ball() {
         this.get_keyboard_response(
             this.record_position.bind(this),
@@ -331,9 +287,6 @@ class Resizer {
         this.ball_animation_frame_id = requestAnimationFrame(this.animate_ball.bind(this));
     }
 
-    /**
-     * Record the ball position.
-     */
     record_position() {
         cancelAnimationFrame(this.ball_animation_frame_id);
         const x = this.accurate_round(this.getElementCenter(this.ball).x, 2);
@@ -352,9 +305,6 @@ class Resizer {
         }
     }
 
-    /**
-     * Finalize the blindspot task.
-     */
     finalize_blindspot_task() {
         const angle = 13.5;
         const sum = this.blindspot_data.ball_pos.reduce((a, b) => a + b, 0);
@@ -394,9 +344,6 @@ class Resizer {
         this.runner._events._currentItem._complete = this._complete_function_cache;
     }
 
-    /**
-     * Reset the ball and wait for the start.
-     */
     reset_ball_wait_for_start() {
         const rectX = this.container.getBoundingClientRect().width - 30;
         const ballX = rectX * 0.85; // define where the ball is
@@ -427,9 +374,6 @@ class Resizer {
         this.ball_animation_frame_id = requestAnimationFrame(this.animate_ball.bind(this));
     }
 
-    /**
-     * adds event listeners to the root element, which allow the keyboard responses to trigger functions.
-     */
     add_root_event_listeners() {
         document.body.addEventListener('keydown', (e) => {
             for (const listener of [...this.listeners]) {
@@ -448,13 +392,6 @@ class Resizer {
         });
     }
 
-    /**
-     * Check if a response is valid.
-     * @param valid_responses array of valid responses
-     * @param allow_held_keys allow held keys
-     * @param key key to check
-     * @returns boolean
-     */	
     check_valid_response(valid_responses: string[], allow_held_keys: boolean = false, key: string) {
         if (!allow_held_keys && this.held_keys.has(key)) {
             return false;
@@ -467,11 +404,6 @@ class Resizer {
         return false;
     }
 
-    /**
-     * Get the center of an element.
-     * @param el element
-     * @returns object with x and y coordinates
-     */	
     getElementCenter(el: HTMLElement) {
         const box = el.getBoundingClientRect();
         return {
@@ -480,9 +412,6 @@ class Resizer {
         };
     }
 
-    /**
-     * removes the event listeners from the root element, which allow the keyboard responses to trigger functions.
-     */ 
     remove_root_event_listeners() {
         this.listeners = [];
         document.body.removeEventListener('keydown', (e) => {
@@ -502,12 +431,6 @@ class Resizer {
         });
     }
 
-    /**
-     * Rounds a number to a specified number of decimal places.
-     * @param value number to round
-     * @param decimals number of decimal places
-     * @returns rounded number
-     */
     accurate_round(value, decimals) {
         return Number(Math.round(Number(value + 'e' + decimals)) + 'e-' + decimals);
     }
