+++ package.json
@@ -12,7 +12,12 @@
     "@jspsych/plugin-html-keyboard-response": "^2.0.0",
     "@jspsych/plugin-virtual-chinrest": "^3.0.0",
     "@types/node": "^22.10.2",
-    "jspsych": "^8.1.0",
-    "typescript": "^5.7.2"
+    "jspsych": "^8.1.0"
+  },
+  "devDependencies": {
+    "@eslint/js": "^9.18.0",
+    "eslint": "^9.18.0",
+    "typescript": "^5.7.3",
+    "typescript-eslint": "^8.20.0"
   }
 }
