--- style.css
@@ -1,3 +1,0 @@
-.resizing-element {
-  border-radius: 50px; /* Adjust the value as needed */
-}
